Remove notification system
When using VS Code in a devcontainer, the theme is not able to 'hot-load' into the environment as it has scripts (specifically, this one) - removing this allows VSC to load our theme into a devcontainer without installing it again.

This should also stop the 'reload required' message when install EC for the first time.